requirements.txt
twisted==16.2.0
pyopenssl==16.0.0
service_identity==16.0.0
enum34==1.1.6
beautifulsoup4==4.4.1
mechanize==0.2.5
requests==2.10.0
hiredis==0.2.0
redis==2.10.5
arrow==0.7.0
sqlalchemy==1.0.13
alembic==0.8.6
psycopg2==2.6.1
lxml==3.6.0
pysocks==1.5.7
wolframalpha==2.4
nltk==3.2.1
markovify==0.4.3
babel==2.3.4
